% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema.R
\name{create_json_schema}
\alias{create_json_schema}
\title{Create a JSON Schema object}
\usage{
create_json_schema(name, schema, strict = TRUE, description = "")
}
\arguments{
\item{name}{Name of the schema}

\item{schema}{The JSON schema definition as a list}

\item{strict}{Whether to enforce strict mode (default TRUE)}

\item{description}{Optional description of the schema}
}
\value{
A json_schema S7 object
}
\description{
Create a JSON Schema object
}
