% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropic_messages.R
\name{.ant_format_schema}
\alias{.ant_format_schema}
\title{Convert json_schema S7 object to Anthropic output_format structure}
\usage{
.ant_format_schema(schema)
}
\description{
Convert json_schema S7 object to Anthropic output_format structure
}
\keyword{internal}
