% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_completions.R
\name{.extract_successful_completion_content}
\alias{.extract_successful_completion_content}
\title{should only pass in successes here, as we're not going to validate types here or try to catch errors and this is \emph{marginally} quicker/memory efficient than pipe + pluck}
\usage{
.extract_successful_completion_content(resp)
}
\description{
should only pass in successes here, as we're not going to validate types here or try to catch errors and this is \emph{marginally} quicker/memory efficient than pipe + pluck
}
\keyword{internal}
