% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndpointR-package.R
\docType{package}
\name{EndpointR-package}
\alias{EndpointR}
\alias{EndpointR-package}
\title{EndpointR: Connects to various Machine Learning inference providers}
\description{
EndpointR is a 'batteries included', open-source R package for connecting to various APIs for Machine Learning model predictions. EndpointR is built for company-specific use cases, so may not be useful to a wide audience.
}
\author{
\strong{Maintainer}: Jack Penzer \email{Jack.penzer@sharecreative.com}

}
\keyword{internal}
