% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_load.R
\name{ep_load}
\alias{ep_load}
\title{Loads Data into the endpointeR}
\usage{
ep_load(file)
}
\arguments{
\item{file}{path and filename as a string (e.g. C:/testfolder/testdata.txt)}
}
\value{
data.frame
}
\description{
Data from a *.txt file will be read into the function so that it can be fed to the endpointR.
Longitudinal data need to be in the wide format. They must have an index column with animal id,
a descriptive column such as treatment. The following columns contain values for consecutive time points. There should be
no missing values until the last measured data point.
}
