% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_select.R
\name{ep_select}
\alias{ep_select}
\title{Animal Selection for endpointR Analysis}
\usage{
ep_select(testdata, n = 1)
}
\arguments{
\item{testdata}{data.frame with test data as provided by \code{ep_load}}

\item{n}{animal index (numeric or character (animal id))}
}
\value{
vector of a single animal for distinct time points.
}
\description{
Using the testdata output from \code{ep_load}, the \code{ep-select} function subsets the testdata by a
user-defined unique animal index. If the input is numeric, e.g. n=1 the first animal in testdata will be chosen. If the
input is a character string, e.g. "ZW21", this specific animal from the testdata will be chosen. Animal ids should not be numeric.
}
\examples{
ep_select(gliodat, n = 1)

}
