% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{epR}
\alias{epR}
\title{Endpointer function}
\usage{
epR(td = td, org = FALSE, wl = 6, SDwdth = 2, mad = FALSE,
  ltype = "b", dotcolor = "black", uprcol = "darkgreen",
  lwrcol = "magenta", cex = 1, cex.axis = 1, cex.lab = 1,
  xlim = NULL, ylim = NULL, pch = 19, blind = FALSE,
  ignupr = FALSE, xlab = "time", ylab = "Moving average (\%)",
  main = NULL)
}
\arguments{
\item{td}{testdata data.frame with n unique rows and p subsequent time points (e.g. days)}

\item{org}{boolean (TRUE/FALSE) for using original values. If FALSE, data are normalized.}

\item{wl}{SD window length (default is 6)}

\item{SDwdth}{width of the standard deviation around the moving average (default is 2.5)}

\item{mad}{boolean - use mean absolute deviation as quasi-clinical scoring constraint (default FALSE)}

\item{ltype}{line type in shown plot}

\item{dotcolor}{color of the shown dots (default "black")}

\item{uprcol}{color of the upper threshold violation indicators}

\item{lwrcol}{color of the upper threshold violation indicators}

\item{cex}{point size}

\item{cex.axis}{axis tick size}

\item{cex.lab}{label size}

\item{xlim}{range of x-axis (if set to NULL (default), plot will adapt automatically to given range - may not be nice!)}

\item{ylim}{range of <-axis (if set to NULL (default), plot will adapt automatically to given range - may not be nice!)}

\item{pch}{R-specific plot symbol for shown dots (default is 19)}

\item{blind}{boolean (TRUE/FALSE) - if set to TRUE, no plot will be shown (default is FALSE)}

\item{ignupr}{boolean (TRUE/FALSE) - ignore upper threshold violations (default is FALSE)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{main}{title}
}
\value{
data.frame with enpointeR results (n=number of data points, timepoint=index of marked endpoint,
where=upper or lower boundary)
}
\description{
The \code{epR} function is a tool for the identification of humane endpoints using single outcome variables
from laboratory animal experiments. Originally, it was developed for using body weight but as values are normalized
other continous variables are suited as well. The algorithm is highly functional in identifying strong deviations from
a windowed normality. The hypothesis behind this is: larger deviations always point to severity.
}
\examples{
epR(as.numeric(gliodat[1,3:length(gliodat[1,])]), blind=TRUE )

}
