% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve an API key which has been stored as an Environment Variable.}
\usage{
get_api_key(key_name)
}
\arguments{
\item{key_name}{The name of the API in the format "ENDPOINT_API_KEY" -> "ANTHROPIC_API_KEY"}
}
\value{
character
}
\description{
Retrieve an API key from .Renviron. The API key should have been set via
\code{set_api_key()}or manually placed in your .Renviron file.
}
\examples{
\dontrun{
  # retrieve an Anthropic API key
  anthropic_key <- get_api_key("ANTHROPIC_API_KEY")

  # use the key in a function call
  my_function(api_key = anthropic_key)
}
}
