% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hf_get_model_max_length}
\alias{hf_get_model_max_length}
\title{Check the max number of tokens allowed for your inputs}
\usage{
hf_get_model_max_length(model_name, api_key = "HF_API_KEY")
}
\arguments{
\item{model_name}{name of the model e.g. 'sentence-transformers/mpnet-base-v2'}

\item{api_key}{Your Hugging Face auth token}
}
\value{
Integer value of the model_max_length from tokenizer config
}
\description{
This function requires the model to have 'tokenizer_config.json' file with a
\code{model_max_length} key, otherwise it will error.
}
