% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_inference.R
\name{hf_perform_request}
\alias{hf_perform_request}
\title{Execute a single embedding request and process the response}
\usage{
hf_perform_request(request, ...)
}
\arguments{
\item{request}{An httr2 request object created by hf_build_request}

\item{...}{ellipsis is sent to \code{httr2::req_perform}, e.g. for \code{path} and \code{verbosity}arguments.}
}
\value{
A httr2 response object
}
\description{
Performs a prepared request and returns the response
}
\examples{
\dontrun{
  # Create and perform request
  req <- hf_build_request(
    input = "This is a sample text to embed",
    endpoint_url = "https://my-endpoint.huggingface.cloud"
  )
  embeddings <- hf_perform_request(req)

}
}
