% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_embed.R
\name{hf_perform_sequential}
\alias{hf_perform_sequential}
\title{Execute multiple embedding requests located in a data frame sequentially}
\usage{
hf_perform_sequential(df, progress = TRUE)
}
\arguments{
\item{df}{A data frame with request objects in a column named '.request'}

\item{progress}{Whether to display a progress bar}
}
\value{
The input data frame with response objects added in '.response' column
}
\description{
Performs multiple prepared embedding requests sequentially and returns the responses.
This function will often be considerably slower than sending requests in parallel.
It is left to the user to decide whether they wish to proceed sequentially or in parallel.
For example, \code{hf_embed_df} has arguments which determine the strategy.
}
\examples{
\dontrun{
  # Prepare requests
  df <- data.frame(
    id = 1:3,
    text = c("First example", "Second example", "Third example")
  )

  requests_df <- hf_build_request_df(
    df = df,
    text_col = "text",
    endpoint_url = "https://my-endpoint.huggingface.cloud"
  )

  # Execute requests sequentially
  responses_df <- hf_perform_sequential(
    df = requests_df,
    progress = TRUE
  )
}
}
