% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_build_embed_req}
\alias{oai_batch_build_embed_req}
\title{Create a single OpenAI Batch API - Embedding request}
\usage{
oai_batch_build_embed_req(
  input,
  id,
  model = "text-embedding-3-small",
  dimensions = NULL,
  method = "POST",
  encoding_format = "float",
  endpoint = "/v1/embeddings"
)
}
\arguments{
\item{input}{Text input you wish to embed}

\item{id}{A custom, unique Row ID}

\item{model}{The embedding model to use}

\item{dimensions}{Number of embedding dimensions to return}

\item{method}{The http request type, usually 'POST'}

\item{encoding_format}{Data type of the embedding values}

\item{endpoint}{The internal suffix of the endpoint's url e.g. /v1/embeddings}
}
\value{
a row of JSON
}
\description{
This function prepares a single row of data for the OpenAI Batch/Files APIs, where each row should be valid JSON. The APIs do not guarantee the results will be in the same order, so we need to provide an ID with each request.
}
