% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_create}
\alias{oai_batch_create}
\title{Trigger a batch job to run on an uploaded file}
\usage{
oai_batch_create(
  file_id,
  endpoint = c("/v1/embeddings", "/v1/chat/completions"),
  completion_window = "24h",
  metadata = NULL,
  key_name = "OPENAI_API_KEY"
)
}
\arguments{
\item{file_id}{Pointer to a file uploaded to the OpenAI API}

\item{endpoint}{The internal suffix of the endpoint's url e.g. /v1/embeddings}

\item{completion_window}{Time until the batch should be returned, NOTE: OpenAI makes 24 hour guarantees only.}

\item{metadata}{Any additional metadata you want to tag the batch with}

\item{key_name}{Name of the API key, usually OPENAI_API_KEY}
}
\value{
Metadata about an OpenAI Batch Job Including the batch ID
}
\description{
Trigger a batch job to run on an uploaded file
}
\details{
Once a file has been uploaded to the OpenAI Files API it's necessary to trigger the batch job. This will ensure that your file is processed, and processing is finalised within the 24 hour guarantee.

It's important to choose the right endpoint. If processing should be done by the Completions API, be sure to route to v1/chat/completions, and this must match each row in your uploaded file.

Batch Job Ids start with "batch_", you'll receive a warning if you try to check batch status on a Files API file (the Files/Batch API set up is a lil bit clumsy for me)
}
