% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_file_upload}
\alias{oai_batch_file_upload}
\title{Prepare and upload a file to be uploaded to the OpenAI Batch API}
\usage{
oai_batch_file_upload(
  jsonl_rows,
  key_name = "OPENAI_API_KEY",
  purpose = "batch"
)
}
\arguments{
\item{jsonl_rows}{Rows of valid JSON, output of a oai_batch_prepare* function}

\item{key_name}{Name of the API key, usually OPENAI_API_KEY}

\item{purpose}{Tag, e.g. 'classification', 'batch', 'fine-tuning'}
}
\value{
Metadata for an upload to the OpenAI Files API
}
\description{
Prepare and upload a file to be uploaded to the OpenAI Batch API
}
\seealso{
\code{openai_files_api.R}
}
