% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_parse_completions}
\alias{oai_batch_parse_completions}
\title{Parse a completions batch job into a data frame}
\usage{
oai_batch_parse_completions(content, original_df = NULL, id_var = NULL)
}
\arguments{
\item{id_var}{}
}
\description{
Parse a completions batch job into a data frame
}
