% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_status}
\alias{oai_batch_status}
\title{Check the status of a batch job on the OpenAI Batch API}
\usage{
oai_batch_status(batch_id, key_name = "OPENAI_API_KEY")
}
\arguments{
\item{batch_id}{Batch Identifier, should start with 'batch_' and is returned by the \code{oai_create_batch} function}

\item{key_name}{Name of the API key, usually OPENAI_API_KEY}
}
\value{
Metadata about an OpenAI Batch API Job, including status, error_file_id, output_file_id, input_file_id etc.
}
\description{
Check the status of a batch job on the OpenAI Batch API
}
