% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_files_api.R
\name{oai_file_content}
\alias{oai_file_content}
\title{Retrieve content from a file on the OpenAI Files API}
\usage{
oai_file_content(file_id, key_name = "OPENAI_API_KEY")
}
\arguments{
\item{file_id}{ID of the file given by OpenAI}

\item{key_name}{The name of your API key, usually "OPENAI_API_KEY"}
}
\description{
Retrieve content from a file on the OpenAI Files API
}
