% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_files_api.R
\name{oai_file_delete}
\alias{oai_file_delete}
\title{Delete a file from the OpenAI Files API}
\usage{
oai_file_delete(file_id, key_name = "OPENAI_API_KEY")
}
\arguments{
\item{file_id}{ID of the file given by OpenAI}

\item{key_name}{The name of your API key, usually "OPENAI_API_KEY"}
}
\description{
Delete a file from the OpenAI Files API
}
