% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema.R
\name{schema_array}
\alias{schema_array}
\title{Create array property schema}
\usage{
schema_array(items, description = NULL, min_items = NULL, max_items = NULL)
}
\arguments{
\item{items}{Schema definition for array elements (use other schema_* helpers)}

\item{description}{Human-readable field description}

\item{min_items}{Minimum number of array elements}

\item{max_items}{Maximum number of array elements}
}
\description{
Defines list/array fields containing multiple items of the same type. Use for
tags, categories, or any collection data. Can constrain array length.
}
\examples{
# List of tags
schema_array(schema_string(), "Product tags", max_items = 10)

# Array of objects
schema_array(
  schema_object(
    name = schema_string("Ingredient name"),
    amount = schema_number("Amount needed")
  ),
  "Recipe ingredients"
)

}
