% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hf_get_endpoint_info}
\alias{hf_get_endpoint_info}
\title{Retrieve information about an endpoint}
\usage{
hf_get_endpoint_info(endpoint_url, key_name = "HF_API_KEY")
}
\arguments{
\item{endpoint_url}{Hugging Face Embedding Endpoint}

\item{key_name}{Name of environment variable containing the API key (default: "HF_API_KEY")}
}
\description{
Retrieve information about an endpoint
}
