% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema.R
\name{json_schema}
\alias{json_schema}
\title{Create JSON Schema S7 class for structured outputs}
\usage{
json_schema(
  name = character(0),
  schema = list(),
  strict = TRUE,
  description = ""
)
}
\description{
S7 class for JSON Schema definitions used in, e.g. OpenAI structured outputs
}
