% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_cancel}
\alias{oai_batch_cancel}
\title{Cancel a running batch job on the OpenAI Batch API}
\usage{
oai_batch_cancel(batch_id, key_name = "OPENAI_API_KEY")
}
\arguments{
\item{key_name}{}
}
\description{
Cancel a running batch job on the OpenAI Batch API
}
