% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_file_upload}
\alias{oai_batch_file_upload}
\title{Prepare and upload a file to be uploaded to the OpenAI Batch API}
\usage{
oai_batch_file_upload(
  jsonl_rows,
  key_name = "OPENAI_API_KEY",
  purpose = "batch"
)
}
\arguments{
\item{jsonl_rows}{Rows of valid JSON, output of an oai_batch_prepare* function}

\item{key_name}{Name of the environment variable containing your API key}

\item{purpose}{File purpose tag, e.g. 'batch', 'fine-tune'}
}
\value{
Metadata for an upload to the OpenAI Files API
}
\description{
Prepare and upload a file to be uploaded to the OpenAI Batch API
}
\examples{
\dontrun{
df <- data.frame(
  id = c("doc_1", "doc_2"),
  text = c("Hello world", "Goodbye world")
)
jsonl_content <- oai_batch_prepare_embeddings(df, text_var = text, id_var = id)
file_info <- oai_batch_file_upload(jsonl_content)
file_info$id # Use this ID to create a batch job
}
}
\seealso{
\code{oai_files_upload()}, \code{oai_files_list()}
}
