% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_prepare_completions}
\alias{oai_batch_prepare_completions}
\title{Title}
\usage{
oai_batch_prepare_completions(
  df,
  text_var,
  id_var,
  model = "gpt-4o-mini",
  system_prompt = NULL,
  temperature = 0,
  max_tokens = 500L,
  schema = NULL,
  method = "POST",
  endpoint = "/v1/chat/completions"
)
}
\arguments{
\item{endpoint}{}
}
\description{
Title
}
