% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_batch_api.R
\name{oai_batch_upload}
\alias{oai_batch_upload}
\title{Prepare and upload a file to be uploaded to the OpenAI Batch API}
\usage{
oai_batch_upload(
  jsonl_rows,
  purpose = c("batch", "fine-tune", "assistants", "vision", "user_data", "evals"),
  key_name = "OPENAI_API_KEY",
  endpoint_url = "https://api.openai.com/v1/files"
)
}
\arguments{
\item{jsonl_rows}{Rows of valid JSON, output of an oai_batch_prepare* function}

\item{purpose}{File purpose tag, e.g. 'batch', 'fine-tune'}

\item{key_name}{Name of the environment variable containing your API key}

\item{endpoint_url}{OpenAI API endpoint URL (default: OpenAI's Files API V1)}
}
\value{
Metadata for an upload to the OpenAI Files API
}
\description{
Prepare and upload a file to be uploaded to the OpenAI Batch API
}
\examples{
\dontrun{
df <- data.frame(
  id = c("doc_1", "doc_2"),
  text = c("Hello world", "Goodbye world")
)
jsonl_content <- oai_batch_prepare_embeddings(df, text_var = text, id_var = id)
file_info <- oai_batch_upload(jsonl_content)
file_info$id # Use this ID to create a batch job
}
}
\seealso{
\code{oai_files_upload()}, \code{oai_files_list()}
}
