% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_files_api.R
\name{oai_file_list}
\alias{oai_file_list}
\title{List files available in the OpenAI Files API}
\usage{
oai_file_list(
  purpose = c("batch", "fine-tune", "assistants", "vision", "user_data", "evals"),
  key_name = "OPENAI_API_KEY"
)
}
\arguments{
\item{purpose}{The intended purpose of the uploaded file, one of "batch", "fine-tune", "assistants", "vision", "user_data", "evals"}

\item{key_name}{The name of your API key, usually "OPENAI_API_KEY"}
}
\description{
List files available in the OpenAI Files API
}
